Create index.html
For this game, create a small area and add a initial color to it. Inside it we will be providing shapes with 'square' as initial shape.
You need to provide at least three different shapes and colors.
Now provide two buttons as -
1. 'Change Color' to change the background color.
2. 'Change Shape' to change the shape.
Clicking on these buttons should change the color background color and shape randomly.